refactor: made mtelib internal
